Move "Bully Luna v.4" into a const
This is an example of how consts can clean up the code a little. If you where to ever change to v.5 in the future it saves you from the easy mistake of only changing it in one place and keeps it just as fast as writing the text in each file.